requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
pytest
requests
python-jose
authlib
starlette
jwt
python-dotenv
oauthlib
itsdangerous
httpx
